PythonCode/requirements.txt
pandas==2.0.3
python-dotenv==1.0.1
openai==1.50.2
